begin merger of uwebqsp and webqsp
