+++ examples/llm_plus_gnn/g_retriever.py
@@ -22,7 +22,7 @@ from tqdm import tqdm
 
 from torch_geometric import seed_everything
 from torch_geometric.data import Dataset
-from torch_geometric.datasets import UpdatedWebQSPDataset, WebQSPDataset
+from torch_geometric.datasets import WebQSPDataset
 from torch_geometric.loader import DataLoader
 from torch_geometric.nn.models import GRetriever
 from torch_geometric.nn.nlp import LLM
@@ -666,9 +666,6 @@ if __name__ == "__main__":
     parser.add_argument(
         "--skip_pretrained_llm_eval", action="store_true",
         help="This flag will skip the evaluation of the pretrained LLM.")
-    parser.add_argument(
-        "--updated_qsp", action="store_true",
-        help="This enables the updated and inproved WebQSP dataloader.")
 
     args = parser.parse_args()
     # check if saved model
@@ -681,7 +678,7 @@ if __name__ == "__main__":
     else:
         retrain = True
     if retrain:
-        dataset = UpdatedWebQSPDataset if args.updated_qsp else None
+        dataset = None
         since = time.time()
         prep_time, dataset, gnn_llm_eval_outs = train(
             since, args.epochs, args.gnn_hidden_channels, args.num_gnn_layers,
@@ -696,8 +693,7 @@ if __name__ == "__main__":
         print("E2E tme minus Prep Time =", e2e_time - prep_time, "seconds")
     else:
         gnn_llm_eval_outs = torch.load("gnn_llm_eval_outs.pt")
-        dataset = WebQSPDataset(
-        ) if not args.updated_qsp else UpdatedWebQSPDataset()
+        dataset = None
     print("Here's a demo showcasing how GNN reduces LLM hallucinations:")
     minimal_demo(gnn_llm_eval_outs, dataset, args.lr, args.epochs,
                  args.batch_size, args.eval_batch_size, get_loss,
