+++ torch_geometric/datasets/web_qsp_dataset.py
@@ -1,12 +1,19 @@
-# Alot of code in this file is based on the original G-Retriever paper
-# url: https://arxiv.org/abs/2402.07630
-from typing import Dict, List, Tuple, no_type_check
+import os
+from itertools import chain
+from typing import Iterator, List, Tuple, no_type_check
 
 import numpy as np
 import torch
 from tqdm import tqdm
 
-from torch_geometric.data import Data, InMemoryDataset
+from torch_geometric.data import (
+    Data,
+    InMemoryDataset,
+    LargeGraphIndexer,
+    TripletLike,
+    get_features_for_triplets_groups,
+)
+from torch_geometric.data.large_graph_indexer import EDGE_RELATION
 from torch_geometric.nn.nlp import SentenceTransformer
 
 try:
@@ -18,12 +25,14 @@ except ImportError:
 
 try:
     from pcst_fast import pcst_fast
+
     WITH_PCST = True
 except ImportError:
     WITH_PCST = False
 
 try:
     import datasets
+
     WITH_DATASETS = True
 except ImportError:
     WITH_DATASETS = False
@@ -38,13 +47,20 @@ def retrieval_via_pcst(
     topk: int = 3,
     topk_e: int = 3,
     cost_e: float = 0.5,
+    save_idx: bool = False,
+    override: bool = False,
 ) -> Tuple[Data, str]:
     c = 0.01
-    if len(textual_nodes) == 0 or len(textual_edges) == 0:
-        desc = textual_nodes.to_csv(index=False) + "\n" + textual_edges.to_csv(
-            index=False, columns=["src", "edge_attr", "dst"])
-        graph = Data(x=graph.x, edge_index=graph.edge_index,
-                     edge_attr=graph.edge_attr, num_nodes=graph.num_nodes)
+    if len(textual_nodes) == 0 or len(textual_edges) == 0 or override:
+        desc = (textual_nodes.to_csv(index=False) +
+                "\n" + textual_edges.to_csv(
+                    index=False, columns=["src", "edge_attr", "dst"]))
+        graph = Data(
+            x=graph.x,
+            edge_index=graph.edge_index,
+            edge_attr=graph.edge_attr,
+            num_nodes=graph.num_nodes,
+        )
         return graph, desc
 
     root = -1  # unrooted
@@ -72,9 +88,8 @@ def retrieval_via_pcst(
             indices = e_prizes == topk_e_values[k]
             value = min((topk_e - k) / sum(indices), last_topk_e_value - c)
             e_prizes[indices] = value
-            last_topk_e_value = value * (1 - c)
-        # reduce the cost of the edges such that at least one edge is selected
-        cost_e = min(cost_e, e_prizes.max().item() * (1 - c / 2))
+            last_topk_e_value = value
+        # cost_e = max(min(cost_e, e_prizes.max().item()-c), 0)
     else:
         e_prizes = torch.zeros(graph.num_edges)
 
@@ -124,8 +139,8 @@ def retrieval_via_pcst(
 
     n = textual_nodes.iloc[selected_nodes]
     e = textual_edges.iloc[selected_edges]
-    desc = n.to_csv(index=False) + "\n" + e.to_csv(
-        index=False, columns=["src", "edge_attr", "dst"])
+    desc = (n.to_csv(index=False) + "\n" +
+            e.to_csv(index=False, columns=["src", "edge_attr", "dst"]))
 
     mapping = {n: i for i, n in enumerate(selected_nodes.tolist())}
 
@@ -137,10 +152,20 @@ def retrieval_via_pcst(
     data = Data(x=x, edge_index=edge_index, edge_attr=edge_attr,
                 num_nodes=len(selected_nodes))
 
+    # HACK Added so that the subset of nodes and edges selected can be tracked
+    if save_idx:
+        data['node_idx'] = np.array(graph.node_idx)[selected_nodes]
+        data['edge_idx'] = np.array(graph.edge_idx)[selected_edges]
+
     return data, desc
 
 
-class WebQSPDataset(InMemoryDataset):
+def preprocess_triplet(triplet: TripletLike) -> TripletLike:
+    h, r, t = triplet
+    return str(h).lower(), str(r), str(t).lower()
+
+
+class UpdatedWebQSPDataset(InMemoryDataset):
     r"""The WebQuestionsSP dataset was released as part of
     “The Value of Semantic Parse Labeling for Knowledge
     Base Question Answering”
@@ -150,21 +175,31 @@ class WebQSPDataset(InMemoryDataset):
     Processing based on "G-Retriever: Retrieval-Augmented Generation
     for Textual Graph Understanding and Question Answering".
     Requires datasets and transformers from HuggingFace.
-
-    Args:
-        root (str): Root directory where the dataset should be saved.
-        force_reload (bool, optional): Whether to re-process the dataset.
-            (default: :obj:`False`)
     """
     def __init__(
         self,
         root: str = "",
         force_reload: bool = False,
+        limit: int = -1,
+        include_pcst: bool = True,
     ) -> None:
-        self._check_dependencies()
+        """Construct a WebQSPDataset.
+
+        Args:
+            root (str, optional): Path to save dataset to. Defaults to "".
+            force_reload (bool, optional): Whether to rerun process step.
+                Defaults to False.
+            limit (int, optional): Construct only the first n samples.
+                Defaults to -1 to construct all samples.
+            include_pcst (bool, optional): Whether to include PCST step
+                (See GRetriever paper). Defaults to True.
+        """
+        self.limit = limit
+        self.include_pcst = include_pcst
         self.device = torch.device(
             "cuda" if torch.cuda.is_available() else "cpu")
         super().__init__(root, None, None, force_reload=force_reload)
+        self._load_raw_data()
         self.load(self.processed_paths[0])
 
     def _check_dependencies(self) -> None:
@@ -182,11 +217,26 @@ class WebQSPDataset(InMemoryDataset):
 
     @property
     def raw_file_names(self) -> List[str]:
-        return []
+        return ["raw_data", "split_idxs"]
 
     @property
     def processed_file_names(self) -> List[str]:
-        return ["list_of_graphs.pt", "pre_filter.pt", "pre_transform.pt"]
+        return [
+            "list_of_graphs.pt",
+            "pre_filter.pt",
+            "pre_transform.pt",
+            "large_graph_indexer",
+        ]
+
+    def _save_raw_data(self) -> None:
+        self.raw_dataset.save_to_disk(self.raw_paths[0])
+        torch.save(self.split_idxs, self.raw_paths[1])
+
+    def _load_raw_data(self) -> None:
+        if not hasattr(self, "raw_dataset"):
+            self.raw_dataset = datasets.load_from_disk(self.raw_paths[0])
+        if not hasattr(self, "split_idxs"):
+            self.split_idxs = torch.load(self.raw_paths[1])
 
     def download(self) -> None:
         dataset = datasets.load_dataset("rmanluo/RoG-webqsp")
@@ -199,58 +249,115 @@ class WebQSPDataset(InMemoryDataset):
             torch.arange(len(dataset["validation"])) + len(dataset["train"]),
             "test":
             torch.arange(len(dataset["test"])) + len(dataset["train"]) +
-            len(dataset["validation"])
+            len(dataset["validation"]),
         }
 
-    def process(self) -> None:
-        self.model = SentenceTransformer(
-            "sentence-transformers/all-roberta-large-v1").to(self.device)
-        self.model.eval()
-        self.questions = [i["question"] for i in self.raw_dataset]
-        list_of_graphs = []
-        # encode questions
+        if self.limit >= 0:
+            self.raw_dataset = self.raw_dataset.select(range(self.limit))
+            self.split_idxs = {
+                "train":
+                torch.arange(self.limit // 2),
+                "val":
+                torch.arange(self.limit // 4) + self.limit // 2,
+                "test":
+                torch.arange(self.limit // 4) + self.limit // 2 +
+                self.limit // 4,
+            }
+        self._save_raw_data()
+
+    def _get_trips(self) -> Iterator[TripletLike]:
+        return chain.from_iterable(
+            iter(ds["graph"]) for ds in self.raw_dataset)
+
+    def _build_graph(self) -> None:
+        trips = self._get_trips()
+        self.indexer: LargeGraphIndexer = LargeGraphIndexer.from_triplets(
+            trips, pre_transform=preprocess_triplet)
+
+        # Nodes:
+        nodes = self.indexer.get_unique_node_features()
+        x = self.model.encode(nodes, batch_size=256)  # type: ignore
+        self.indexer.add_node_feature(new_feature_name="x", new_feature_vals=x)
+
+        # Edges:
+        edges = self.indexer.get_unique_edge_features(
+            feature_name=EDGE_RELATION)
+        edge_attr = self.model.encode(edges, batch_size=256)  # type: ignore
+        self.indexer.add_edge_feature(
+            new_feature_name="edge_attr",
+            new_feature_vals=edge_attr,
+            map_from_feature=EDGE_RELATION,
+        )
+
+        print("Saving graph...")
+        self.indexer.save(self.processed_paths[-1])
+
+    def _retrieve_subgraphs(self) -> None:
         print("Encoding questions...")
+        self.questions = [str(ds["question"]) for ds in self.raw_dataset]
         q_embs = self.model.encode(self.questions, batch_size=256)
-        print("Encoding graphs...")
+        list_of_graphs = []
+        print("Retrieving subgraphs...")
+        textual_nodes = self.textual_nodes
+        textual_edges = self.textual_edges
+        if self.whole_graph_retrieval:
+            graph = self.indexer.to_data(node_feature_name="x",
+                                         edge_feature_name="edge_attr")
+        else:
+            graph_gen = get_features_for_triplets_groups(
+                self.indexer, (ds['graph'] for ds in self.raw_dataset),
+                pre_transform=preprocess_triplet)
+
         for index in tqdm(range(len(self.raw_dataset))):
             data_i = self.raw_dataset[index]
-            raw_nodes: Dict[str, int] = {}
-            raw_edges = []
-            for tri in data_i["graph"]:
-                h, r, t = tri
-                h = h.lower()
-                t = t.lower()
-                if h not in raw_nodes:
-                    raw_nodes[h] = len(raw_nodes)
-                if t not in raw_nodes:
-                    raw_nodes[t] = len(raw_nodes)
-                raw_edges.append({
-                    "src": raw_nodes[h],
-                    "edge_attr": r,
-                    "dst": raw_nodes[t]
-                })
-            nodes = DataFrame([{
-                "node_id": v,
-                "node_attr": k
-            } for k, v in raw_nodes.items()], columns=["node_id", "node_attr"])
-            edges = DataFrame(raw_edges, columns=["src", "edge_attr", "dst"])
-            # encode nodes
-            nodes.node_attr = nodes.node_attr.fillna("")
-            x = self.model.encode(nodes.node_attr.tolist(), batch_size=256)
-            # encode edges
-            edge_attr = self.model.encode(edges.edge_attr.tolist(),
-                                          batch_size=256)
-            edge_index = torch.LongTensor(
-                [edges.src.tolist(), edges.dst.tolist()])
+            if not self.whole_graph_retrieval:
+                graph = next(graph_gen)
+                textual_nodes = self.textual_nodes.iloc[
+                    graph["node_idx"]].reset_index()
+                textual_edges = self.textual_edges.iloc[
+                    graph["edge_idx"]].reset_index()
+            pcst_subgraph, desc = retrieval_via_pcst(
+                graph,
+                q_embs[index],
+                textual_nodes,
+                textual_edges,
+                topk=3,
+                topk_e=5,
+                cost_e=0.5,
+                override=not self.include_pcst,
+            )
             question = f"Question: {data_i['question']}\nAnswer: "
             label = ("|").join(data_i["answer"]).lower()
-            raw_graph = Data(x=x, edge_index=edge_index, edge_attr=edge_attr,
-                             num_nodes=len(nodes)).to("cpu")
-            pcst_subgraph, desc = retrieval_via_pcst(raw_graph, q_embs[index],
-                                                     nodes, edges, topk=3,
-                                                     topk_e=5, cost_e=0.5)
+
             pcst_subgraph["question"] = question
             pcst_subgraph["label"] = label
             pcst_subgraph["desc"] = desc
             list_of_graphs.append(pcst_subgraph.to("cpu"))
+        print("Saving subgraphs...")
         self.save(list_of_graphs, self.processed_paths[0])
+
+    def process(self) -> None:
+        self._load_raw_data()
+        self.model = SentenceTransformer(
+            'sentence-transformers/all-roberta-large-v1').to(self.device)
+        self.model.eval()
+        if not os.path.exists(self.processed_paths[-1]):
+            print("Encoding graph...")
+            self._build_graph()
+        else:
+            print("Loading graph...")
+            self.indexer = LargeGraphIndexer.from_disk(
+                self.processed_paths[-1])
+        self.textual_nodes = DataFrame.from_dict(
+            {"node_attr": self.indexer.get_node_features()})
+        self.textual_nodes["node_id"] = self.textual_nodes.index
+        self.textual_nodes = self.textual_nodes[["node_id", "node_attr"]]
+        self.textual_edges = DataFrame(self.indexer.get_edge_features(),
+                                       columns=["src", "edge_attr", "dst"])
+        self.textual_edges["src"] = [
+            self.indexer._nodes[h] for h in self.textual_edges["src"]
+        ]
+        self.textual_edges["dst"] = [
+            self.indexer._nodes[h] for h in self.textual_edges["dst"]
+        ]
+        self._retrieve_subgraphs()
